requirements.txt
# Core Flask dependencies
flask>=3.1.0
flask-sqlalchemy>=3.1.1
flask-migrate>=4.1.0
flask-login>=0.6.3
flask-wtf>=1.2.2
flask-mail>=0.10.0
flask-cors>=4.0.0
flask-caching>=2.1.0

# Task scheduling for cache maintenance
schedule>=1.2.0

# Database and ORM
sqlalchemy>=2.0.40
psycopg2-binary>=2.9.10
alembic>=1.13.0

# Security and utilities
werkzeug>=3.1.3
wtforms>=3.2.1
email-validator>=2.2.0
python-dotenv>=1.1.0
pyotp>=2.9.0
markupsafe>=2.1.0
user-agents>=2.2.0

# HTTP and external requests
requests>=2.31.0
urllib3>=2.0.0

# File handling and security
python-magic>=0.4.27

# AWS S3 for file storage
boto3>=1.35.0
botocore>=1.35.0

# Image processing for QR codes
qrcode>=7.4.2
pillow>=10.0.0
pytz>=2023.3

# System monitoring and performance
psutil>=5.9.0

# Production server
gunicorn>=23.0.0